+++ src/main/kotlin/com/kotlin/sns/domain/Posting/service/Impl/PostingServiceImpl.kt
@@ -1,14 +1,15 @@
 package com.kotlin.sns.domain.Posting.service.Impl
 
+import com.kotlin.sns.common.exception.CustomException
+import com.kotlin.sns.common.exception.ExceptionConst
 import com.kotlin.sns.domain.Posting.dto.request.RequestCreaetePostingDto
 import com.kotlin.sns.domain.Posting.dto.request.RequestUpdatePostingDto
 import com.kotlin.sns.domain.Posting.dto.response.ResponsePostingDto
-import com.kotlin.sns.domain.Posting.entity.Posting
 import com.kotlin.sns.domain.Posting.mapper.PostingMapper
 import com.kotlin.sns.domain.Posting.repository.PostingRepository
 import com.kotlin.sns.domain.Posting.service.PostingService
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
-import java.lang.IllegalArgumentException
 import kotlin.reflect.full.memberProperties
 
 /**
@@ -24,14 +25,20 @@ class PostingServiceImpl(
 ) : PostingService {
 
     /**
-     * uuid기반으로 반환
+     * uuid 기반으로 posting 반환
      *
      * @param postingId
      * @return
      */
-    override fun findPostingById(postingId : Long) : ResponsePostingDto{
+    override fun findPostingById(postingId: Long): ResponsePostingDto {
         val posting = postingRepository.findById(postingId)
-            .orElseThrow { IllegalArgumentException("invalid posting id : $postingId") }
+            .orElseThrow {
+                CustomException(
+                    ExceptionConst.POSTING,
+                    HttpStatus.NOT_FOUND,
+                    "Posting with id $postingId not found"
+                )
+            }
 
         return postingMapper.toDto(posting)
     }
@@ -57,16 +64,22 @@ class PostingServiceImpl(
     override fun updatePosting(requestUpdatePostingDto: RequestUpdatePostingDto): ResponsePostingDto {
         val postingId = requestUpdatePostingDto.postingId
         val posting = postingRepository.findById(postingId)
-            .orElseThrow { IllegalArgumentException("invalid posting id : $postingId") }
+            .orElseThrow {
+                CustomException(
+                    ExceptionConst.POSTING,
+                    HttpStatus.NOT_FOUND,
+                    "Posting with id $postingId not found"
+                )
+            }
 
         val fields = requestUpdatePostingDto::class.memberProperties
 
         for (field in fields) {
-            val fieldName = field.name  //필드의 이름
-            val fieldValue = field.getter.call(requestUpdatePostingDto)//필드가 가진 value
+            val fieldName = field.name  // 필드의 이름
+            val fieldValue = field.getter.call(requestUpdatePostingDto)  // 필드의 값
 
-            if(fieldValue != null){
-                when(fieldName){
+            if (fieldValue != null) {
+                when (fieldName) {
                     "content" -> posting.content = fieldValue as String
                     "imageUrl" -> posting.imageUrl = fieldValue as String?
                 }
@@ -82,8 +95,13 @@ class PostingServiceImpl(
      * @param postingId
      */
     override fun deletePosting(postingId: Long) {
+        if (!postingRepository.existsById(postingId)) {
+            throw CustomException(
+                ExceptionConst.POSTING,
+                HttpStatus.NOT_FOUND,
+                "Posting with id $postingId not found"
+            )
+        }
         postingRepository.deleteById(postingId)
     }
-
-
 }
